move commands to internal
