+++ cmd/internal/base/mux_test.go
@@ -1,40 +1,40 @@
-package main_test
+package base_test
 
 import (
 	"testing"
 
-	"ariga.io/atlas/cmd/atlas"
+	"ariga.io/atlas/cmd/internal/base"
 	"github.com/stretchr/testify/require"
 )
 
 func Test_ProviderNotSupported(t *testing.T) {
-	u := main.NewMux()
+	u := base.NewMux()
 	_, err := u.OpenAtlas("fake://open")
 	require.Error(t, err)
 }
 
 func Test_RegisterProvider(t *testing.T) {
-	u := main.NewMux()
-	p := func(s string) (*main.Driver, error) { return nil, nil }
+	u := base.NewMux()
+	p := func(s string) (*base.Driver, error) { return nil, nil }
 	require.NotPanics(t, func() { u.RegisterProvider("key", p) })
 }
 
 func Test_RegisterTwiceSameKeyFails(t *testing.T) {
-	u := main.NewMux()
-	p := func(s string) (*main.Driver, error) { return nil, nil }
+	u := base.NewMux()
+	p := func(s string) (*base.Driver, error) { return nil, nil }
 	require.NotPanics(t, func() { u.RegisterProvider("key", p) })
 	require.Panics(t, func() { u.RegisterProvider("key", p) })
 }
 
 func Test_GetDriverFails(t *testing.T) {
-	u := main.NewMux()
+	u := base.NewMux()
 	_, err := u.OpenAtlas("key://open")
 	require.Error(t, err)
 }
 
 func Test_GetDriverSuccess(t *testing.T) {
-	u := main.NewMux()
-	p := func(s string) (*main.Driver, error) { return nil, nil }
+	u := base.NewMux()
+	p := func(s string) (*base.Driver, error) { return nil, nil }
 	u.RegisterProvider("key", p)
 	_, err := u.OpenAtlas("key://open")
 	require.NoError(t, err)
