+++ cmd/internal/base/provider.go
@@ -1,4 +1,4 @@
-package main
+package base
 
 import (
 	"database/sql"
