+++ go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/go-sql-driver/mysql v1.6.0
 	github.com/google/go-cmp v0.5.6 // indirect
 	github.com/hashicorp/hcl/v2 v2.10.0
-	github.com/jackc/pgconn v1.10.0
 	github.com/kr/pretty v0.2.0 // indirect
 	github.com/lib/pq v1.10.3
 	github.com/manifoldco/promptui v0.9.0
@@ -15,6 +14,7 @@ require (
 	github.com/stretchr/testify v1.7.0
 	github.com/zclconf/go-cty v1.8.0
 	golang.org/x/mod v0.4.3-0.20210512182355-6088ed88cecd
+	golang.org/x/sys v0.0.0-20210615035016-665e8c7367d1 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
